+++ app/client.jsx
@@ -1,28 +1,22 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
-import { Router } from 'react-router';
-import { syncReduxAndRouter } from 'redux-simple-router';
-import createBrowserHistory from 'history/lib/createBrowserHistory';
-import routes from 'routes.jsx';
+import { Router, browserHistory } from 'react-router';
+import createRoutes from 'routes.jsx';
 import configureStore from 'store/configureStore';
 
 // Grab the state from a global injected into
 // server-generated HTML
 const initialState = window.__INITIAL_STATE__;
 
-const store = configureStore(initialState);
-const history = createBrowserHistory();
-
-// Installs hooks that always keep react-router and redux
-// store in sync
-syncReduxAndRouter(history, store);
+const store = configureStore(initialState, browserHistory);
+const routes = createRoutes(store);
 
 // Router converts <Route> element hierarchy to a route config:
 // Read more https://github.com/rackt/react-router/blob/latest/docs/Glossary.md#routeconfig
 render(
   <Provider store={store}>
-    <Router history={history}>
+    <Router history={browserHistory}>
       {routes}
     </Router>
   </Provider>, document.getElementById('app'));