+++ app/routes.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Route from 'react-router';
+import { Route, IndexRoute } from 'react-router';
 
 import App from 'containers/App';
 import Vote from 'containers/Vote';
@@ -7,13 +7,28 @@ import About from 'containers/About';
 import Login from 'containers/Login';
 import Dashboard from 'containers/Dashboard';
 
-import { requireAuthentication } from 'components/authenticateComponent';
-
-export default (
-  <Route component={App}>
-    <Route path="/" component={Vote} />
-    <Route path="login" component={Login} />
-    <Route path="dashboard" component={requireAuthentication(Dashboard)} />
-    <Route path="about" component={About} />
-  </Route>
-);
+/*
+ * @param {Redux Store}
+ * We require store as an argument here because we wish to get
+ * state from the store after it has been authenticated.
+ */
+export default (store) => {
+  const requireAuth = (nextState, replace, callback) => {
+    const { user: { authenticated }} = store.getState();
+    if (!authenticated) {
+      replace({
+        pathname: '/login',
+        state: { nextPathname: nextState.location.pathname }
+      });
+    }
+    callback();
+  };
+  return (
+    <Route path="/" component={App}>
+      <IndexRoute component={Vote} />
+      <Route path="login" component={Login} />
+      <Route path="dashboard" component={Dashboard} onEnter={requireAuth} />
+      <Route path="about" component={About} />
+    </Route>
+  );
+};