+++ app/server.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import { renderToString } from 'react-dom/server';
-import { RoutingContext, match } from 'react-router'
-import createLocation from 'history/lib/createLocation';
+import { RouterContext, match, createMemoryHistory } from 'react-router'
 import fetch from 'isomorphic-fetch';
 import { Provider } from 'react-redux';
-import routes from 'routes.jsx';
+import createRoutes from 'routes.jsx';
 import configureStore from 'store/configureStore';
 import headconfig from 'elements/Header';
 
@@ -65,41 +64,64 @@ function renderFullPage(renderedContent, initialState, head={
  */
 export default function render(req, res) {
 
-  // Note that req.url here should be the full URL path from
-  // the original request, including the query string.
-  match({ routes, location: req.url }, (error, redirectLocation, renderProps) => {
-    if (error) {
-      res.status(500).send(error.message);
-    } else if (redirectLocation) {
-      res.redirect(302, redirectLocation.pathname + redirectLocation.search);
-    } else if (renderProps) {
-      fetchTopics(apiResult => {
-        const authenticated = req.isAuthenticated();
-        const store = configureStore({
-          // reducer: {initialState}
-          topic: {
-            topics: apiResult
-          },
-          user: {
-            authenticated: authenticated, 
-            isWaiting: false
-          }
-        });
+  fetchTopics(apiResult => {
+    const history = createMemoryHistory();
+    const authenticated = req.isAuthenticated();
+    const store = configureStore({
+      // reducer: {initialState}
+      topic: {
+        topics: apiResult
+      },
+      user: {
+        authenticated: authenticated, 
+        isWaiting: false
+      }
+    }, history);
+    const routes = createRoutes(store);
+
+    /*
+     * From the react-router docs:
+     * 
+     * This function is to be used for server-side rendering. It matches a set of routes to
+     * a location, without rendering, and calls a callback(error, redirectLocation, renderProps)
+     * when it's done.
+     *
+     * The function will create a `history` for you, passing additional `options` to create it.
+     * These options can include `basename` to control the base name for URLs, as well as the pair
+     * of `parseQueryString` and `stringifyQuery` to control query string parsing and serializing.
+     * You can also pass in an already instantiated `history` object, which can be constructured 
+     * however you like.
+     *
+     * The three arguments to the callback function you pass to `match` are:
+     * - error: A javascript Error object if an error occured, `undefined` otherwise.
+     * - redirectLocation: A `Location` object if the route is a redirect, `undefined` otherwise
+     * - renderProps: The props you should pass to the routing context if the route matched, `undefined`
+     *                otherwise.
+     * If all three parameters are `undefined`, this means that there was no route found matching the
+     * given location.
+     */
+    match({ routes, location: req.url }, (error, redirectLocation, renderProps) => {
+      if (error) {
+        res.status(500).send(error.message);
+      } else if (redirectLocation) {
+        res.redirect(302, redirectLocation.pathname + redirectLocation.search);
+      } else if (renderProps) {
+
         const initialState = store.getState();
         const renderedContent = renderToString(
         <Provider store={store}>
-          <RoutingContext {...renderProps} />
+          <RouterContext {...renderProps} />
         </Provider>);
         const renderedPage = renderFullPage(renderedContent, initialState, {
           title: headconfig.title,
           meta: headconfig.meta,
           link: headconfig.link
         });
-        res.status(200).send(renderedPage);
-      });
-    } else {
-      res.status(404).send('Not Found');
-    }
-    
+        res.status(200).send(renderedPage);              
+      } else {
+        res.status(404).send('Not Found');
+      }
+
+    });
   });
 };